Create Simple Interset
In this program, we first get the principal amount, time, and rate from the user using the input() function. 
We convert these inputs to float values using the float() function, as we need to perform arithmetic operations on them.
Then, we calculate the simple interest using the formula (PxTxR)/100, and store the result in the simple_interest variable.
Finally, we print the result using the print() function. The result will be displayed in the format "Simple Interest = x", where x is the calculated simple interest.