Create Python program to find smallest number in a list  Using function
In this program, we define a function called find_smallest_number() that takes a list as an argument.
Inside the function, we initialize the smallest variable with the first element of the list. Then, we loop through the list using a for loop and compare each element with the smallest variable. If an element is smaller than smallest, we update the value of smallest.
Finally, we return the smallest variable from the function.
To test the function, we create a list list1, call the find_smallest_number() function with this list as an argument, and print the result. The output will be the smallest number in the list, which in this case is 4