Create Python Program for factorial of a number using function, loop and if else conditions.    For example factorial of 5 is 5*4*3*2*1 which is 120
In this program, we define a function called factorial() that takes a number as an argument. Inside the function, we first define a variable fact to store the factorial of the number.Next, we check if the number is 0 or 1. If it is, we return 1 as the factorial of 0 or 1 is 1.If the number is not 0 or 1, we loop through the range from 1 to the number using a for loop. In each iteration, we multiply the value of fact by the current value of i. At the end of the loop, the value of fact will be the factorial of the input number.Finally, we return the value of fact from the function.To test the function, we create a variable input_num with the input number (in this case, 5), call the factorial() function with input_num as an argument, and print the result. The output will be the factorial of the input number, which in this case is 120.